PYTHON-6012 - Consolidate TopologySettings into a single shared class
